Update: PAT-252 - Support DBAL connection
