+++ composer.json
@@ -6,7 +6,8 @@
         "PHP": "^7.4|^8",
         "keboola/csv": "^1.1",
         "tracy/tracy": "^2.3",
-        "aws/aws-sdk-php": "^3.11"
+        "aws/aws-sdk-php": "^3.11",
+        "doctrine/dbal": "^3.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.5",
